Simplify the base app even further by removing unnecessary code
In particular:
- we don't need local notifications
- we don't need referrals
- we don't want to make any remote calls whatsoever (after https://github.com/e-mission/e-mission-phone/pull/470/commits/27376007b85f6ebe7f6d50220610284687195530) since the final server is not configured yet and the user has not logged in.

So we can delete all those files and their module references in the retained files.

Second, we don't need to make any remote calls, we can also remove the code
that sets the connection settings on startup. The server won't be the correct
one anyway; that will be configured as part of the channel that we connect to.

So the startup code gets enormously simplified as well.