Bump up version numbers
The only change in this app is the fix for https://github.com/e-mission/cordova-jwt-auth/issues/33 merged from https://github.com/e-mission/cordova-jwt-auth/pull/34

So the only changes are to bump up the version number for the auth plugin AND
to bump up the version number for the app to reflect that there has been a
change.

Pretty simple!